# openapi.yaml
openapi: 3.1.0
info:
  title: Draftmark API
  description: |
    Markdown sharing platform for async collaboration between humans and AI agents.

    The core loop: agent/human writes markdown → shares link → reviewer reads + comments → agent consumes feedback via API.

    ## Authentication

    Draftmark uses three token types — no passwords, no OAuth:

    | Token | Prefix | Passed via | Purpose |
    |-------|--------|------------|---------|
    | Magic Token | _(none)_ | `X-Magic-Token` header or `?token=` query | Owner operations (update, delete) |
    | Doc API Key | `key_` | `Authorization: Bearer key_…` | Read private docs, manage comments |
    | Account API Key | `acct_` | `Authorization: Bearer acct_…` | Account-level operations |

    All tokens are shown **once** at creation time and stored as SHA-256 hashes.

    ## Review Lifecycle

    Documents accept feedback while `status` is `open` **and** the `review_deadline` (if set) has not passed.
    Three mechanisms control this:

    1. **Explicit close** — PATCH with `status: "review_closed"`
    2. **Threshold** — Set `expected_reviews`; `review_complete` becomes `true` when met (signal only)
    3. **Deadline** — Set `review_deadline`; feedback is auto-rejected (409) after expiry

    ## Cross-Session Context

    Agents should write a `.draftmark.json` file after creating a doc:
    ```json
    { "slug": "abc12345", "api_key": "key_…", "url": "https://draftmark.app/share/abc12345" }
    ```
  version: 1.0.0
...
servers:
...
paths:
...
  /docs:
    post:
      operationId: createDoc
      summary: Create a document
      tags: [Documents]
      security:
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocRequest"
      responses:
        "201":
          description: Document created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateDocResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /docs/{slug}:
    parameters:
      - $ref: "#/components/parameters/Slug"

    get:
      operationId: getDoc
      summary: Get a document
      description: |
        Returns document content and metadata. For private docs, requires auth via
        Bearer API key, magic token, or account key.

        Use `?format=raw` to get raw markdown as `text/markdown`.
        Use `?view=executive|technical|ux` for stakeholder-specific views (if available).

        Shortcut: `GET /share/{slug}.md` returns raw markdown directly (via middleware rewrite).
      tags: [Documents]
      security:
        - {}
        - BearerApiKey: []
        - MagicTokenHeader: []
        - MagicTokenQuery: []
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [raw]
          description: Set to `raw` to return markdown as `text/markdown`
        - name: view
          in: query
          schema:
            type: string
            enum: [executive, technical, ux]
          description: Return a stakeholder-specific view (if available in metadata)
        - name: token
          in: query
          schema:
            type: string
          description: Magic token for owner access
        - name: share_token
          in: query
          schema:
            type: string
          description: Read-only share token for private doc access (share_ prefix)
      responses:
        "200":
          description: Document with metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doc"
            text/markdown:
              schema:
                type: string
                description: Raw markdown content (when format=raw)
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      operationId: updateDoc
      summary: Update a document
      description: Requires magic token or account ownership. Updates content, status, review settings, or visibility.
      tags: [Documents]
      security:
        - MagicTokenHeader: []
        - MagicTokenQuery: []
        - BearerAccountKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocRequest"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateDocResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteDoc
      summary: Delete a document
      tags: [Documents]
      security:
        - MagicTokenHeader: []
        - MagicTokenQuery: []
        - BearerAccountKey: []
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Comments ───────────────────────────────────────────────
...
components:
...
  securitySchemes:
    BearerApiKey:
      type: http
      scheme: bearer
      description: "Doc-level API key (`key_` prefix)"

    BearerAccountKey:
      type: http
      scheme: bearer
      description: "Account API key (`acct_` prefix)"

    MagicTokenHeader:
      type: apiKey
      in: header
      name: X-Magic-Token
      description: Magic token for owner operations

    MagicTokenQuery:
      type: apiKey
      in: query
      name: token
      description: Magic token as query parameter

  # ── Parameters ─────────────────────────────────────────────
  parameters:
    Slug:
      name: slug
      in: path
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 8
      description: "8-character nanoid slug"

  # ── Reusable Responses ─────────────────────────────────────
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    DeleteResponse:
      type: object
      required: [deleted]
      properties:
        deleted:
          type: boolean
          example: true

    # ── Document schemas ───────────────────────────────────
    CreateDocRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          description: Markdown content
        title:
          type: string
          description: Auto-extracted from first heading if omitted
        visibility:
          type: string
          enum: [public, private]
          default: public
        version_note:
          type: string
        expected_reviews:
          type: integer
          minimum: 1
          description: Number of reviews to trigger `review_complete` flag
        review_deadline:
          type: string
          format: date-time
          description: ISO 8601 date after which feedback is auto-rejected
        meta:
          type: object
          description: Arbitrary JSON metadata
    CreateDocResponse:
      type: object
      required: [slug, url, magic_token, api_key]
      properties:
        slug:
          type: string
          example: "a1b2c3d4"
        url:
          type: string
          format: uri
          example: "https://draftmark.app/share/a1b2c3d4"
        magic_token:
          type: string
          description: "Owner token. Shown once — store it."
        api_key:
          type: string
          description: "Doc API key (key_ prefix). Shown once — store it."
        share_token:
          type: string
          description: "Read-only share token (share_ prefix). Only returned for private docs. Retrievable via GET /docs/:slug with owner auth."
        share_url:
          type: string
          format: uri
          description: "Shareable URL with embedded share token. Only returned for private docs."
          example: "https://draftmark.app/share/a1b2c3d4?share_token=share_abc123"
    Doc:
      type: object
      required:
        - slug
        - content
        - visibility
        - status
        - accepting_feedback
        - reactions_count
        - comments_count
        - reviews_count
        - created_at
        - updated_at
      properties:
        slug:
          type: string
        title:
          type: [string, "null"]
        content:
          type: string
        available_views:
          type: array
          items:
            type: string
          description: Stakeholder view names available (e.g. executive, technical, ux)
        visibility:
          type: string
          enum: [public, private]
        status:
          type: string
          enum: [open, review_closed]
        expected_reviews:
          type: [integer, "null"]
        review_deadline:
          type: [string, "null"]
          format: date-time
        review_complete:
          type: boolean
          description: True when reviews_count >= expected_reviews
        review_expired:
          type: boolean
          description: True when review_deadline has passed
        accepting_feedback:
          type: boolean
          description: True when status is open and deadline has not passed
        reactions_count:
          $ref: "#/components/schemas/ReactionCounts"
        comments_count:
          type: integer
        reviews_count:
          type: integer
        reviews:
          type: array
          items:
            type: object
            properties:
              reviewer_name:
                type: string
              reviewed_at:
                type: string
                format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        # Owner-only fields (require magic_token or account ownership)
        views_count:
          type: integer
          description: "Owner-only. Total page views."
        meta:
          type: object
          description: "Owner-only. Arbitrary JSON metadata."
        share_token:
          type: string
          description: "Owner-only. Read-only share token for private docs."
        share_url:
          type: string
          format: uri
          description: "Owner-only. Shareable URL with embedded share token for private docs."
    UpdateDocRequest:
      type: object
      properties:
        content:
          type: string
        title:
          type: string
        visibility:
          type: string
          enum: [public, private]
        version_note:
          type: string
          description: Only used when content is also provided
        status:
          type: string
          enum: [open, review_closed]
        expected_reviews:
          type: [integer, "null"]
          minimum: 1
          description: Set to null to clear
        review_deadline:
          type: [string, "null"]
          format: date-time
          description: Set to null to clear
    UpdateDocResponse:
      type: object
      required: [slug, content, visibility, status, updated_at]
      properties:
        slug:
          type: string
        title:
          type: [string, "null"]
        content:
          type: string
        visibility:
          type: string
          enum: [public, private]
        status:
          type: string
          enum: [open, review_closed]
        expected_reviews:
          type: [integer, "null"]
        review_deadline:
          type: [string, "null"]
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ReactionCounts:
      type: object
      properties:
        thumbs_up:
          type: integer
          default: 0
        check:
          type: integer
          default: 0
        thinking:
          type: integer
          default: 0
        cross:
          type: integer
          default: 0
...